@@ -9,7 +9,6 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "eslint": "^8.47.0",
     "react": "^18.0.0",
     "react-dom": "^18.0.0"
   },
@@ -20,6 +19,7 @@
     "@types/react-dom": "^18.0.0",
     "@vitejs/plugin-react": "^1.3.0",
     "typescript": "^4.6.3",
-    "vite": "^2.9.9"
+    "vite": "^2.9.9",
+    "eslint": "^9.25.1"
   }
 }
