@@ -4,6 +4,7 @@
   "private": true,
   "packageManager": "pnpm@8.6.12",
   "license": "MIT",
+  "type": "module",
   "scripts": {
     "lint": "eslint . --fix",
     "test": "pnpm -r run test",
@@ -12,14 +13,16 @@
     "prepare": "husky"
   },
   "devDependencies": {
-    "@cmkk/prettier-config": "workspace:*",
-    "@cmkk/eslint-config": "workspace:*",
     "@cmkk/commitlint-config": "workspace:*",
+    "@cmkk/eslint-config": "workspace:*",
+    "@cmkk/prettier-config": "workspace:*",
+    "@eslint/eslintrc": "^3.3.1",
+    "@eslint/js": "^9.25.1",
     "bumpp": "^9.1.1",
-    "eslint": "^8.46.0",
+    "eslint": "^9.25.1",
+    "husky": "^9.1.7",
     "prettier": "^3.0.3",
     "rimraf": "^5.0.1",
-    "husky": "^9.1.7",
     "typescript": "^4.9.5"
   },
   "directories": {
@@ -34,5 +37,7 @@
   "bugs": {
     "url": "https://github.com/rankangkang/config/issues"
   },
-  "homepage": "https://github.com/rankangkang/config#readme"
+  "homepage": "https://github.com/rankangkang/config#readme",
+  "dependencies": {
+  }
 }
